# Decode bytes versus runes.